Add support for tvOS
This was tested with prebuilt TVVLCKit only. There is no compatibility
with custom VLCKit builds.

When VLCMediaLibraryKit was compiled both for iOS and tvOS, a
xcframework including binaries for both OS will be created, otherwise
just for the OS the library was compiled for.

The xcframework is no longer created by default. The newly introduced -f
flag must be used.